Add Segment calculations for extended segments
A segment might need to be extended one cell further in each direction
e.g. for the outer edge of a tank. Or it might need to exclude the last
cell, so that 4 edges of 4 cells each can be concatenated to form the
boundary of a tank (or the points near a base where a tank can move to
in order to destroy the base).