Add two-valued circular buffer to do SSSP calc's
I've thought of a way to do breadth first search by designing edge
weights in the graph to all have weight 1 and having a node per: cell +
direction + action type (move/fire). The circular buffer allows
efficient implementation of a queue for the BFS algorithm. Enhance it to
track two sets of distances: at insertion point and at removal point,
with a pointer to the cell where the insertion point values are higher
than the removal value.