CacheNode best as a struct (but Node is faster)
Modified DistanceCalculator and DistanceCalculation to use CacheNode
instead of Node. This ran very slowly with CacheNode as a class. It ran
much faster with CacheNode as a struct.

It was still fastest with the Node class though, which is basically just
an embellished int!

Reverted the DistanceCalculation and DistanceCalculator code before
committing.