Added attack calculations FROM a tank
Previously attack calculations were to a single target point,
and movement calculations were from a single source point.
This performs an attack calculation from a single source point to a
single target point, using the movement calculations for the single
source point, and the firing line calculations for the target point.